Define module-level fixtures for component tests
