+++ tests/conftest.py
@@ -7,12 +7,8 @@ from unittest import mock
 
 import pytest
 
-from kpops.component_handlers import ComponentHandlers
-from kpops.component_handlers.helm_wrapper.model import HelmDiffConfig
-from kpops.config import KpopsConfig, TopicNameConfig, set_config
 from kpops.utils.environment import ENV, Environment
 from kpops.utils.yaml import load_yaml_file
-from tests.components import PIPELINE_BASE_DIR
 
 logger = logging.getLogger("faker")
 logger.setLevel(logging.INFO)  # quiet faker locale messages
@@ -55,26 +51,3 @@ def custom_components():
         yield
     finally:
         shutil.rmtree(dst)
-
-
-@pytest.fixture(autouse=True, scope="session")
-def config() -> None:
-    config = KpopsConfig(
-        topic_name_config=TopicNameConfig(
-            default_error_topic_name="${component.type}-error-topic",
-            default_output_topic_name="${component.type}-output-topic",
-        ),
-        kafka_brokers="broker:9092",
-        helm_diff_config=HelmDiffConfig(),
-        pipeline_base_dir=PIPELINE_BASE_DIR,
-    )
-    set_config(config)
-
-
-@pytest.fixture(autouse=True, scope="session")
-def handlers() -> ComponentHandlers:
-    return ComponentHandlers(
-        schema_handler=mock.AsyncMock(),
-        connector_handler=mock.AsyncMock(),
-        topic_handler=mock.AsyncMock(),
-    )
